% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation.R
\name{dGPD_QRM}
\alias{dGPD_QRM}
\title{This function is the dGPD function from the QRM package.}
\usage{
dGPD_QRM(x, xi, beta = 1, log = FALSE)
}
\arguments{
\item{x}{Vector. The data used to estimate GPD parameters from.}

\item{xi}{Numeric. The xi parameter from the GPD distribution.}

\item{beta}{Numeric. The beta parameter from the GPD distribution.}

\item{log}{Logical. If TRUE, will not return the log of the estimate.}
}
\value{
Vector of length \code{length(x)}. The density of the GPD evaluated at
each \code{x} value.
}
\description{
This function is the dGPD function from the QRM package.
}
\examples{

x <- 1:10; xi <- 0.5;

dGPD_QRM(x = x, xi = xi)
}
