% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EVTr-package.R
\docType{package}
\name{EVTr}
\alias{EVTr}
\title{Estimating censored long term injuries with Extreme Value Theory.}
\description{
Estimating censored long term injuries with Extreme Value Theory
}
\details{
\pkg{EVTr} is a package that accompanies a masters thesis called “Using extreme value
theory to model severe injuries of circus artists”.

Important functions:

\itemize{
\item \code{\link{gen_data}}: Simulate injuries according to a generalized Pareto distribution and
healthy periods according to a exponential distribution. Based on a fixed parameter \code{censor},
some of the injuries and healthy periods will be censored based on the cummulative sum of the
healthy and injury periods.

\item \code{\link{simulation}}: function that generates data according to \code{specific}
method and then applies the maximum likelihood estimator (mle) based on \code{method}.

\item \code{\link{mle}}: function to estimate the maximum likelihood estimator for either
the regular generalized Pareto distribution or the censored generalized Pareto distribution.

\item \code{\link{excess}}: calculate the excess above a given threshold based on \code{ne}.
The threshold is calculated so that there are exactly \code{ne} observations above it.

}

See individual function documentation for usage.
}
\author{
Annabelle Redelmeier
}
