% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{findthreshold_QRM}
\alias{findthreshold_QRM}
\title{Calculate the threshold such that there are \code{ne} instances from \code{data} above this threshold.
This code comes from the QRM package.}
\usage{
findthreshold_QRM(data, ne)
}
\arguments{
\item{data}{Dataframe. Contains at least the columns "ID", "Injury_Length", and "Censored".}

\item{ne}{Integer. Count of excesses above the threshold.}
}
\value{
Dataframe. Of observations above the threshold (decided by \code{ne}) and the excess above the threshold.
}
\description{
Calculate the threshold such that there are \code{ne} instances from \code{data} above this threshold.
This code comes from the QRM package.
}
\examples{
data <- rexp(10)

findthreshold_QRM(data, ne = 1)

}
