% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_data.R
\name{gen_data}
\alias{gen_data}
\title{Simulate injuries according to a generalized Pareto distribution and healthy periods according to a exponential
distribution.
Based on a fixed parameter \code{censor}, some of the injuries and healthy periods will be censored based on the
cummulative sum of the healthy and injury periods. Based on \code{specific}, certain observations will be returned.}
\usage{
gen_data(
  censor = NULL,
  xi,
  n,
  num_inj,
  rate_exp,
  ne = NULL,
  specific = c("delete_censored_obs", "keep_censored_obs", "keep_only_max_obs",
    "max_excess", "excess"),
  seed = 1
)
}
\arguments{
\item{censor}{Integer. Where to censor observations.}

\item{xi}{Numeric. The xi parameter in the generalized Pareto distribution.}

\item{n}{Integer. The number of individuals.}

\item{num_inj}{Integer. The number of injuries per individual.}

\item{rate_exp}{The rate in the exponential distribution.}

\item{ne}{Integer. Number of observations above the threshold. Only used when \code{specific} = 'max_excess' or
'excess'.}

\item{specific}{String. Corresponding to one of "delete_censored_obs", "keep_censored_obs", "keep_only_max_obs",
"max_excess" or "excess".}

\item{seed}{Integer.}
}
\value{
Dataframe.

This dataframe will have various columns depending on \code{specific}.

If \code{specific} = "delete_censored_obs", the first column in \code{data_return} is "Injury_Length" and corresponds
to the length of the injury.

The second column "ID" stands for the invididuals ID or identifier. "Censored" = 1 if the
injury is censored before the entire injury is finished (in which case "Injury_Length" is the length of the
injury until the censoring) and = 0 otherwise.

"Actual" indicates the length of the injury before it was censored. In other words, if there was no censoring,
we would see the actual length of the injury.

"Any_Injury_Censored" == 1 if the individual has had their last injury censored. Since all individuals are censored
at one point, each individual while either have an injury censored or a healthy period censored. But since we don't
care about modelling healthy periods, if the healthy period is censored, we will have full information on the
injuries.

"Prop_of_injuries_censored" is the # injuries censored / # individuals (n).

"Prop_of_ind_with_censored_injury" is the # injuries censored / total # injuries.

"Prop_max_censored" is the # individuals where the censored injury is their longest injury / # individuals.

If \code{specific} = "keep_censored_obs", the columns are the same.

If \code{specific} = "keep_only_max_obs", the above columns all apply. A new column "delta" is:
1 - (1 - Censored) * Any_Injury_Censored

If \code{specific} = "max_excess", the above columns all apply.  In this dataframe, "Injury_Length" is the excess
injury length above the threshold decided by \code{ne}. A new column "Injury_Length_before" is the original length
of the injury (i.e threshold - Injury_Length).

"Prop_of_obs_above_threshold" = # injuries >= threshold / total # of injuries.

If \code{specific} = "excess", the columns are the same as "max_excess".
}
\description{
Simulate injuries according to a generalized Pareto distribution and healthy periods according to a exponential
distribution.
Based on a fixed parameter \code{censor}, some of the injuries and healthy periods will be censored based on the
cummulative sum of the healthy and injury periods. Based on \code{specific}, certain observations will be returned.
}
\details{
\code{delete_censored_obs}: simulates data, censors the observations, but only returns the fully observed (injury)
periods.
\code{keep_censored_obs}: simulates data, censors the observations, and returns all (injury) periods.
\code{keep_only_max_obs}: simulates data, censors, returns only the longest injury (censored or not censored)
for each individual.
\code{max_excess}: simulates data, censors, calculates the longest injury (censored or not censored), then
calculates the threshold based on \code{ne} (number of exceedances above the threshold). Finally returns only
the excess above this given threshold.
\code{excess}: simulates data, censors, then calculates the threshold based on \code{ne} (number of exceedances above
the threshold). Finally returns only the excess above this given threshold.
}
\examples{
censor <- 3; xi <- 0.5; n <- 2; num_inj <- 3; rate_exp <- 1; ne <- 1

gen_data(censor = censor, xi = xi, n = n, num_inj = num_inj, rate_exp = rate_exp,
specific = "delete_censored_obs")
}
\author{
Annabelle Redelmeier
}
