% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation.R
\name{mle_gpd}
\alias{mle_gpd}
\title{A function that calculates the maximum likelihood estimator based on the generalized Pareto distribution.}
\usage{
mle_gpd(data)
}
\arguments{
\item{data}{Dataframe. One column is called "Injury_Length" and corresponds to the observations.}
}
\value{
Numeric. Corresponding to the maximum likelihood estimator based on the generalized Pareto distribution.
}
\description{
A function that calculates the maximum likelihood estimator based on the generalized Pareto distribution.
}
\examples{

data <- data.frame(Injury_Length = rexp(10))

mle_gpd(data = data)
}
\author{
Annabelle Redelmeier
}
