% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation.R
\name{mle_gpd_full}
\alias{mle_gpd_full}
\title{This function actually comes from the R package QRM and is only used to test to see if the other mle
functions work properly.}
\usage{
mle_gpd_full(data, threshold, information = c("observed", "expected"))
}
\arguments{
\item{data}{Dataframe. One column is called "Injury_Length" and corresponds to the observations.}

\item{threshold}{Numeric. Threshold to use when using the peaks-over-threshold method.}

\item{information}{String. One of "observed" or "expected" used in optimization.}
}
\value{
List. Contains  "n" which is the number of observations in \code{data},
"p.less.thresh" which is the proportion below the threshold,
"n.exceed" is the number of observations above the threshold, "par.ests" is a vector
containing the parameter estimates of xi and beta, "par.ses" contains the standard errors
of the estimated parameters, and "varcov" contains the covariance matrix of the estimated
parameters.
}
\description{
This function actually comes from the R package QRM and is only used to test to see if the other mle
functions work properly.
}
\examples{

data <- data.frame(Injury_Length = rexp(100))

mle_gpd_full(data = data, threshold = 1, information = "expected")
}
