% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation.R
\name{negloglik_cens_gpd_full}
\alias{negloglik_cens_gpd_full}
\title{The negative censored log-likelihood of the censored generalized Pareto distribution implemented by the \code{dGPD} R package.}
\usage{
negloglik_cens_gpd_full(theta, ydata)
}
\arguments{
\item{theta}{Vector. Length 2. The first argument corresponds to $$xi$$ in the GPD distribution and the second
argument corresponds to $$beta$$.}

\item{ydata}{Dataframe. the first column corresponding to the data/observations. The second column corresponding to the
binary column. 0 if the observation is not censored. 1 if the observation is censored.}
}
\value{
Numeric. Corresponding to the maximum likelihood estimator based on the generalized Pareto distribution.
}
\description{
The negative censored log-likelihood of the censored generalized Pareto distribution implemented by the \code{dGPD} R package.
}
\examples{

data <- data.frame(obs = rexp(10), Censored = rbinom(10, 1, 0.5))

negloglik_cens_gpd_full(theta = c(1, 1), ydata = data)
}
\author{
Annabelle Redelmeier
}
