% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation.R
\name{negloglik_gpd}
\alias{negloglik_gpd}
\title{The negative log-likelihood of the generalized Pareto distribution.}
\usage{
negloglik_gpd(theta, tmp)
}
\arguments{
\item{theta}{Vector. Length 2. The first argument corresponds to $$xi$$ in the GPD distribution and the second
argument corresponds to $$beta$$.}

\item{tmp}{Dataframe. One column corresponding to the data/observations.}
}
\value{
Numeric. Corresponding to the maximum likelihood estimator based on the generalized Pareto distribution.
}
\description{
The negative log-likelihood of the generalized Pareto distribution.
}
\examples{

data <- rexp(10)

negloglik_gpd(theta = c(1, 1), tmp = data)
}
\author{
Annabelle Redelmeier
}
