% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation.R
\name{negloglik_gpd_full}
\alias{negloglik_gpd_full}
\title{The negative log-likelihood of the generalized Pareto distribution implemented in the \code{QRM} R package.}
\usage{
negloglik_gpd_full(theta, ydata)
}
\arguments{
\item{theta}{Vector. Length 2. The first argument corresponds to $$xi$$ in the GPD distribution and the second
argument corresponds to $$beta$$.}

\item{ydata}{Dataframe. One column corresponding to the data/observations.}
}
\value{
Numeric. Corresponding to the maximum likelihood estimator based on the generalized Pareto distribution.
}
\description{
The negative log-likelihood of the generalized Pareto distribution implemented in the \code{QRM} R package.
}
\examples{

data <- rexp(10)

negloglik_gpd_full(theta = c(1, 1), ydata = data)
}
\author{
Annabelle Redelmeier
}
