% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_data.R
\name{rGPD_QRM}
\alias{rGPD_QRM}
\title{This is the rGPD function in the QRM package. It is used to generate generalized Pareto distribution
random variables.}
\usage{
rGPD_QRM(n, xi, beta = 1)
}
\arguments{
\item{n}{Integer. The number of random variables to generate.}

\item{xi}{Numeric. The xi parameter in the generalized Pareto distribution.}

\item{beta}{Numeric. The beta parameter in the generalized Pareto distribution.}
}
\value{
Vector. Length equal to \code{n}. Corresponds to random variables from the GPD distribution.
}
\description{
This is the rGPD function in the QRM package. It is used to generate generalized Pareto distribution
random variables.
}
\examples{

n <- 10; xi <- 1;
rGPD_QRM(n, xi)

}
