% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_data.R
\name{simulation}
\alias{simulation}
\title{This function generates data according to \code{specific} method and then applies the maximum likelihood estimator (mle)
based on \code{method}.}
\usage{
simulation(
  censor = NULL,
  xi,
  n,
  num_inj,
  rate_exp,
  ne = NULL,
  specific,
  seed,
  method
)
}
\arguments{
\item{censor}{Integer. Where to censor observations.}

\item{xi}{Numeric. The xi parameter in the generalized Pareto distribution.}

\item{n}{Integer. The number of individuals.}

\item{num_inj}{Integer. The number of injuries per individual.}

\item{rate_exp}{The rate in the exponential distribution.}

\item{ne}{Integer. Number of observations above the threshold. Only used when \code{specific} = 'max_excess' or
'excess'.}

\item{specific}{String. Corresponding to one of "delete_censored_obs", "keep_censored_obs", "keep_only_max_obs",
"max_excess" or "excess".}

\item{seed}{Numeric. Seed to set.}

\item{method}{String. Corresponding to one of ""MLE", "CensMLE", "MLE_delete_censored", "N", "MLE_full" or "CensMLE_full".}
}
\value{
Numeric. Corresponding to the maximum likelihood estimator based on \code{method}.
}
\description{
This function generates data according to \code{specific} method and then applies the maximum likelihood estimator (mle)
based on \code{method}.
}
\examples{

censor <- 3; xi <- 0.5; n <- 10; num_inj <- 3; rate_exp <- 1
specific <- "delete_censored_obs"
seed <- 1
method <- "MLE"

simulation(censor = censor, xi = xi, n = n, num_inj = num_inj, rate_exp = rate_exp,
specific = "delete_censored_obs",
seed = seed, method = method)[]
}
\author{
Annabelle Redelmeier
}
