% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation.R
\name{mle}
\alias{mle}
\title{A function to estimate the maximum likelihood estimator for a variety of methods.}
\usage{
mle(data, method = c("MLE", "CensMLE"))
}
\arguments{
\item{data}{Dataframe. One column is called "Injury_Length" and corresponds to the observations.}

\item{method}{String. One of "MLE", "CensMLE", "MLE_full", or "CensMLE_full".}
}
\value{
Numeric. Corresponding to the maximum likelihood estimator based on one of the generalized Pareto distribution.
}
\description{
A function to estimate the maximum likelihood estimator for a variety of methods.
}
\examples{

data <- data.frame(Injury_Length = rexp(10), Censored = rbinom(10, 1, 0.5))

mle(data = data, method = "MLE")
}
\author{
Annabelle Redelmeier
}
