% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation.R
\name{mle_gpd_full}
\alias{mle_gpd_full}
\title{A function that calculates the maximum likelihood estimator based on the generalized Pareto distribution.}
\usage{
mle_gpd_full(
  data,
  threshold,
  information = c("observed", "expected"),
  verbose = TRUE,
  ...
)
}
\arguments{
\item{data}{Dataframe. One column is called "Injury_Length" and corresponds to the observations.}

\item{threshold}{Numeric. Corresponds to the threshold. Only the excess above the threshold will be used to
estimate the parameters of the generalized Pareto distribution.}

\item{information}{String. Either "observed" or "expected".}

\item{verbose.}{Logical. If TRUE, prints more information.}
}
\value{
Numeric. Corresponding to the maximum likelihood estimator based on the generalized Pareto distribution.
}
\description{
A function that calculates the maximum likelihood estimator based on the generalized Pareto distribution.
}
\examples{

data <- data.frame(Injury_Length = rexp(10))

mle_gpd_full(data = data, threshold = 0.5, information = "observed", verbose = TRUE)
}
\author{
Annabelle Redelmeier
}
