% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation.R
\name{negloglik_cens_gpd}
\alias{negloglik_cens_gpd}
\title{The negative censored log-likelihood of the generalized Pareto distribution.}
\usage{
negloglik_cens_gpd(theta, tmp)
}
\arguments{
\item{theta}{Vector. Length 2. The first argument corresponds to the xi parameter in the GPD distribution and
the second argument corresponds to the beta parameter.}

\item{tmp}{Dataframe. Two columns. First column are the observations. The second column is binary indicating
which of the observations are censored.}
}
\value{
Numeric. Corresponding to the maximum likelihood estimator based on the censored generalized Pareto
distribution.
}
\description{
The negative censored log-likelihood of the generalized Pareto distribution.
}
\examples{

data <- rexp(10)
data <- cbind(data, censored = rbinom(10, 1, 0.5))

negloglik_cens_gpd(theta = c(1, 1), tmp = data)
}
\author{
Annabelle Redelmeier
}
